+++ Finance/Finance/Data/HttpRequestFactory.cs
@@ -2,7 +2,7 @@ namespace Finance.Data;
 
 public class HttpRequestFactory : IHttpRequestFactory 
 {
-    public HttpRequestMessage GetYahooFinanceRequest(string stock, double start, double end) =>
+    public HttpRequestMessage GetYahooFinanceRequest(string stock, long start, long end) =>
         new HttpRequestMessage
         {
             Method = HttpMethod.Get,
@@ -14,11 +14,11 @@ public class HttpRequestFactory : IHttpRequestFactory
             }
         };
 
-    private string GetGetYahooFinanceQueryString(string stock, double start, double end) =>
+    private string GetGetYahooFinanceQueryString(string stock, long start, long end) =>
         $"?symbol={stock}&from={start}&to={end}&events=div&interval=1d&region=UK";
 }
 
 public interface IHttpRequestFactory
 {
-    HttpRequestMessage GetYahooFinanceRequest(string stock, double start, double end);
+    HttpRequestMessage GetYahooFinanceRequest(string stock, long start, long end);
 }
